FIXING ALL SILLY ERRORS, which were a lot

+++ src/main/java/com/pluralsight/IPriceable/IPriceable.java
@@ -1,9 +1,9 @@
 package com.pluralsight.IPriceable;
 
-import java.math.BigDecimal;
+
 
 public interface IPriceable {
-    BigDecimal calculatePrice();
+    double calculatePrice();
 }
